Add safety check to update_prod.

+++ update_prod.sh
@@ -1,12 +1,16 @@
 export PGPASSWORD=ymctas
 
-sudo echo "working"
+read -p "dropping soccerstats db - are you sure?" -n 1 -r
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  echo "rpdating db"
+  dropdb soccerstats
+  psql -U soccerstats -d soccerstats_dev -c "CREATE DATABASE soccerstats WITH TEMPLATE soccerstats_dev"
 
-dropdb soccerstats
-psql -U soccerstats -d soccerstats_dev -c "CREATE DATABASE soccerstats WITH TEMPLATE soccerstats_dev"
+  echo "restarting site"
+  sudo stop s2prod
+  sudo start s2prod
+  sudo /etc/init.d/memcached restart
 
-sudo stop s2prod
-sudo start s2prod
-sudo /etc/init.d/memcached restart
-
-python manage.py rebuild_index
+  python manage.py rebuild_index
+fi
